+++ sdk/storage/azure-storage-file-datalake/src/test/java/com/azure/storage/file/datalake/ServiceAPITest.groovy
@@ -10,18 +10,17 @@ import com.azure.identity.DefaultAzureCredentialBuilder
 import com.azure.storage.blob.BlobUrlParts
 import com.azure.storage.blob.models.BlobStorageException
 import com.azure.storage.common.ParallelTransferOptions
-import com.azure.storage.common.sas.AccountSasPermission
-import com.azure.storage.common.sas.AccountSasResourceType
-import com.azure.storage.common.sas.AccountSasService
-import com.azure.storage.common.sas.AccountSasSignatureValues
-import com.azure.storage.file.datalake.implementation.util.DataLakeImplUtils
-import com.azure.storage.file.datalake.models.DataLakeRequestConditions
+import com.azure.storage.file.datalake.models.DataLakeAnalyticsLogging
+import com.azure.storage.file.datalake.models.DataLakeCorsRule
+import com.azure.storage.file.datalake.models.DataLakeMetrics
+import com.azure.storage.file.datalake.models.DataLakeRetentionPolicy
+import com.azure.storage.file.datalake.models.DataLakeServiceProperties
+import com.azure.storage.file.datalake.models.DataLakeStaticWebsite
 import com.azure.storage.file.datalake.models.DataLakeStorageException
 import com.azure.storage.file.datalake.models.FileSystemItem
 import com.azure.storage.file.datalake.models.FileSystemListDetails
 import com.azure.storage.file.datalake.models.ListFileSystemsOptions
 import com.azure.storage.file.datalake.models.UserDelegationKey
-import com.azure.storage.file.datalake.options.FileSystemRenameOptions
 import com.azure.storage.file.datalake.options.FileSystemUndeleteOptions
 import reactor.core.publisher.Mono
 import reactor.test.StepVerifier
@@ -30,6 +29,178 @@ import java.time.Duration
 import java.time.OffsetDateTime
 
 class ServiceAPITest extends APISpec {
+
+    def validatePropsSet(DataLakeServiceProperties sent, DataLakeServiceProperties received) {
+        return received.getLogging().isRead() == sent.getLogging().isRead() &&
+            received.getLogging().isDelete() == sent.getLogging().isDelete() &&
+            received.getLogging().isWrite() == sent.getLogging().isWrite() &&
+            received.getLogging().getVersion() == sent.getLogging().getVersion() &&
+            received.getLogging().getRetentionPolicy().getDays() == sent.getLogging().getRetentionPolicy().getDays() &&
+            received.getLogging().getRetentionPolicy().isEnabled() == sent.getLogging().getRetentionPolicy().isEnabled() &&
+
+            received.getCors().size() == sent.getCors().size() &&
+            received.getCors().get(0).getAllowedMethods() == sent.getCors().get(0).getAllowedMethods() &&
+            received.getCors().get(0).getAllowedHeaders() == sent.getCors().get(0).getAllowedHeaders() &&
+            received.getCors().get(0).getAllowedOrigins() == sent.getCors().get(0).getAllowedOrigins() &&
+            received.getCors().get(0).getExposedHeaders() == sent.getCors().get(0).getExposedHeaders() &&
+            received.getCors().get(0).getMaxAgeInSeconds() == sent.getCors().get(0).getMaxAgeInSeconds() &&
+
+            received.getDefaultServiceVersion() == sent.getDefaultServiceVersion() &&
+
+            received.getHourMetrics().isEnabled() == sent.getHourMetrics().isEnabled() &&
+            received.getHourMetrics().isIncludeApis() == sent.getHourMetrics().isIncludeApis() &&
+            received.getHourMetrics().getRetentionPolicy().isEnabled() == sent.getHourMetrics().getRetentionPolicy().isEnabled() &&
+            received.getHourMetrics().getRetentionPolicy().getDays() == sent.getHourMetrics().getRetentionPolicy().getDays() &&
+            received.getHourMetrics().getVersion() == sent.getHourMetrics().getVersion() &&
+
+            received.getMinuteMetrics().isEnabled() == sent.getMinuteMetrics().isEnabled() &&
+            received.getMinuteMetrics().isIncludeApis() == sent.getMinuteMetrics().isIncludeApis() &&
+            received.getMinuteMetrics().getRetentionPolicy().isEnabled() == sent.getMinuteMetrics().getRetentionPolicy().isEnabled() &&
+            received.getMinuteMetrics().getRetentionPolicy().getDays() == sent.getMinuteMetrics().getRetentionPolicy().getDays() &&
+            received.getMinuteMetrics().getVersion() == sent.getMinuteMetrics().getVersion() &&
+
+            received.getDeleteRetentionPolicy().isEnabled() == sent.getDeleteRetentionPolicy().isEnabled() &&
+            received.getDeleteRetentionPolicy().getDays() == sent.getDeleteRetentionPolicy().getDays() &&
+
+            received.getStaticWebsite().isEnabled() == sent.getStaticWebsite().isEnabled() &&
+            received.getStaticWebsite().getIndexDocument() == sent.getStaticWebsite().getIndexDocument() &&
+            received.getStaticWebsite().getErrorDocument404Path() == sent.getStaticWebsite().getErrorDocument404Path()
+    }
+
+    def "Set get properties"() {
+        when:
+        def retentionPolicy = new DataLakeRetentionPolicy().setDays(5).setEnabled(true)
+        def logging = new DataLakeAnalyticsLogging().setRead(true).setVersion("1.0")
+            .setRetentionPolicy(retentionPolicy)
+        def corsRules = new ArrayList<DataLakeCorsRule>()
+        corsRules.add(new DataLakeCorsRule().setAllowedMethods("GET,PUT,HEAD")
+            .setAllowedOrigins("*")
+            .setAllowedHeaders("x-ms-version")
+            .setExposedHeaders("x-ms-client-request-id")
+            .setMaxAgeInSeconds(10))
+        def defaultServiceVersion = "2016-05-31"
+        def hourMetrics = new DataLakeMetrics().setEnabled(true).setVersion("1.0")
+            .setRetentionPolicy(retentionPolicy).setIncludeApis(true)
+        def minuteMetrics = new DataLakeMetrics().setEnabled(true).setVersion("1.0")
+            .setRetentionPolicy(retentionPolicy).setIncludeApis(true)
+        def website = new DataLakeStaticWebsite().setEnabled(true)
+            .setIndexDocument("myIndex.html")
+            .setErrorDocument404Path("custom/error/path.html")
+
+        def sentProperties = new DataLakeServiceProperties()
+            .setLogging(logging).setCors(corsRules).setDefaultServiceVersion(defaultServiceVersion)
+            .setMinuteMetrics(minuteMetrics).setHourMetrics(hourMetrics)
+            .setDeleteRetentionPolicy(retentionPolicy)
+            .setStaticWebsite(website)
+
+        def headers = primaryDataLakeServiceClient.setPropertiesWithResponse(sentProperties, null, null).getHeaders()
+
+        // Service properties may take up to 30s to take effect. If they weren't already in place, wait.
+        sleepIfRecord(30 * 1000)
+
+        def receivedProperties = primaryDataLakeServiceClient.getProperties()
+
+        then:
+        headers.getValue("x-ms-request-id") != null
+        headers.getValue("x-ms-version") != null
+        validatePropsSet(sentProperties, receivedProperties)
+    }
+
+    // In java, we don't have support from the validator for checking the bounds on days. The service will catch these.
+
+    def "Set props min"() {
+        setup:
+        def retentionPolicy = new DataLakeRetentionPolicy().setDays(5).setEnabled(true)
+        def logging = new DataLakeAnalyticsLogging().setRead(true).setVersion("1.0")
+            .setRetentionPolicy(retentionPolicy)
+        def corsRules = new ArrayList<DataLakeCorsRule>()
+        corsRules.add(new DataLakeCorsRule().setAllowedMethods("GET,PUT,HEAD")
+            .setAllowedOrigins("*")
+            .setAllowedHeaders("x-ms-version")
+            .setExposedHeaders("x-ms-client-request-id")
+            .setMaxAgeInSeconds(10))
+        def defaultServiceVersion = "2016-05-31"
+        def hourMetrics = new DataLakeMetrics().setEnabled(true).setVersion("1.0")
+            .setRetentionPolicy(retentionPolicy).setIncludeApis(true)
+        def minuteMetrics = new DataLakeMetrics().setEnabled(true).setVersion("1.0")
+            .setRetentionPolicy(retentionPolicy).setIncludeApis(true)
+        def website = new DataLakeStaticWebsite().setEnabled(true)
+            .setIndexDocument("myIndex.html")
+            .setErrorDocument404Path("custom/error/path.html")
+
+
+        def sentProperties = new DataLakeServiceProperties()
+            .setLogging(logging).setCors(corsRules).setDefaultServiceVersion(defaultServiceVersion)
+            .setMinuteMetrics(minuteMetrics).setHourMetrics(hourMetrics)
+            .setDeleteRetentionPolicy(retentionPolicy)
+            .setStaticWebsite(website)
+
+        expect:
+        primaryDataLakeServiceClient.setPropertiesWithResponse(sentProperties, null, null).getStatusCode() == 202
+    }
+
+    def "Set props cors check"() {
+        setup:
+        def serviceProperties = primaryDataLakeServiceClient.getProperties()
+
+        // Some properties are not set and this test validates that they are not null when sent to the service
+        def rule = new DataLakeCorsRule()
+        rule.setAllowedOrigins("microsoft.com")
+        rule.setMaxAgeInSeconds(60)
+        rule.setAllowedMethods("GET")
+        rule.setAllowedHeaders("x-ms-version")
+
+        serviceProperties.setCors(Collections.singletonList(rule))
+
+        expect:
+        primaryDataLakeServiceClient.setPropertiesWithResponse(serviceProperties, null, null).getStatusCode() == 202
+    }
+
+    def "Set props static website"() {
+        setup:
+        def serviceProperties = primaryDataLakeServiceClient.getProperties()
+        def errorDocument404Path = "error/404.html"
+        def defaultIndexDocumentPath = "index.html"
+
+        serviceProperties.setStaticWebsite(new DataLakeStaticWebsite()
+            .setEnabled(true)
+            .setErrorDocument404Path(errorDocument404Path)
+            .setDefaultIndexDocumentPath(defaultIndexDocumentPath)
+        )
+
+        when:
+        Response<Void> resp = primaryDataLakeServiceClient.setPropertiesWithResponse(serviceProperties, null, null)
+
+        then:
+        resp.getStatusCode() == 202
+        def staticWebsite = primaryDataLakeServiceClient.getProperties().getStaticWebsite()
+        staticWebsite.isEnabled()
+        staticWebsite.getErrorDocument404Path() == errorDocument404Path
+        staticWebsite.getDefaultIndexDocumentPath() == defaultIndexDocumentPath
+    }
+
+    def "Set props error"() {
+        when:
+        getServiceClient(primaryCredential, "https://error.blob.core.windows.net")
+            .setProperties(new DataLakeServiceProperties())
+
+        then:
+        thrown(DataLakeStorageException)
+    }
+
+    def "Get props min"() {
+        expect:
+        primaryDataLakeServiceClient.getPropertiesWithResponse(null, null).getStatusCode() == 200
+    }
+
+    def "Get props error"() {
+        when:
+        getServiceClient(primaryCredential, "https://error.blob.core.windows.net")
+            .getProperties()
+
+        then:
+        thrown(DataLakeStorageException)
+    }
 
     def "List file systems"() {
         when: